Refactor directory structure
- refactor directory structure
  - convertor > convertors
  - generator > generators
- fix npm command
  - artifact script bug-fix
