+++ tsconfig.json
@@ -35,9 +35,9 @@
     "isolatedModules": true,
     "baseUrl": "." /* Base directory to resolve non-absolute module names. */,
     "paths": {
-      "#convertor/*": ["src/convertor/*"],
+      "#convertors/*": ["src/convertors/*"],
       "#interfaces/*": ["src/interfaces/*"],
-      "#generator/*": ["src/generator/*"],
+      "#generators/*": ["src/generators/*"],
       "#tools/*": ["src/tools/*"]
     } /* A series of entries which re-map imports to lookup locations relative to the 'baseUrl'. */,
     // "typeRoots": ["node_modules/@types", "src/typings"] /* List of folders to include type definitions from. */ ,
